+++ school-medical-management/src/main/java/com/swp391/school_medical_management/modules/users/dtos/request/NotifyToParentRequest.java
@@ -1,9 +1,10 @@
 package com.swp391.school_medical_management.modules.users.dtos.request;
 
-import java.util.List;
 import lombok.Data;
+
+import java.util.List;
 
 @Data
 public class NotifyToParentRequest {
-   private List<Long> formIds; 
+    private List<Integer> formIds;
 }
