+++ school-medical-management/src/main/java/com/swp391/school_medical_management/modules/users/repositories/MedicalRecordsRepository.java
@@ -1,14 +1,12 @@
 package com.swp391.school_medical_management.modules.users.repositories;
 
-import java.util.List;
-import java.util.Optional;
-
+import com.swp391.school_medical_management.modules.users.entities.MedicalRecordEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.swp391.school_medical_management.modules.users.entities.MedicalRecordEntity;
+import java.util.Optional;
 
-public interface MedicalRecordsRepository extends JpaRepository<MedicalRecordEntity, Long> {
-     Optional<MedicalRecordEntity> findByStudent_Id(int studentId);
+public interface MedicalRecordsRepository extends JpaRepository<MedicalRecordEntity, Integer> {
+    Optional<MedicalRecordEntity> findByStudent_Id(int studentId);
     // List<MedicalRecordEntity> findMedicalRecordByStudent_Parent_UserId(Long parentId);
     // long count();
